Moved sealed class for scanning setup UI events
