+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/scanningsetup/WooPosScanningSetupUiEvent.kt
@@ -1,0 +1,10 @@
+package com.woocommerce.android.ui.woopos.home.scanningsetup
+
+import com.woocommerce.android.ui.woopos.home.scanningsetup.WooPosScanningSetupState.BarcodeReaderDevice
+
+sealed class WooPosScanningSetupUiEvent {
+    data object OnPrimaryButtonClicked : WooPosScanningSetupUiEvent()
+    data object OnSecondaryButtonClicked : WooPosScanningSetupUiEvent()
+    data object OnOpenBluetoothSettings : WooPosScanningSetupUiEvent()
+    data class OnDeviceSelected(val device: BarcodeReaderDevice) : WooPosScanningSetupUiEvent()
+}
